Refactor: Move entity classes to models package.
