+++ src/main/java/edu/ucsd/sbrg/escher/models/Element.java
@@ -14,44 +14,28 @@
  * <http://www.gnu.org/licenses/lgpl-3.0-standalone.html>.
  * ---------------------------------------------------------------------
  */
-package edu.ucsd.sbrg.escher;
+package edu.ucsd.sbrg.escher.models;
 
 /**
  * @author Andreas Dr&auml;ger
  *
  */
-public interface Position extends EscherBase {
+public interface Element extends EscherBase {
 
 	/**
-	 * @return the x
+	 * @return the id
 	 */
-	public Double getX();
-
-	/**
-	 * @return the y
-	 */
-	public Double getY();
+	public String getId();
 
 	/**
 	 * 
 	 * @return {@code true} if the requested property is not {@code null}.
 	 */
-	public boolean isSetX();
+	public boolean isSetId();
 
 	/**
-	 * 
-	 * @return {@code true} if the requested property is not {@code null}.
+	 * @param id the id to set
 	 */
-	public boolean isSetY();
-
-	/**
-	 * @param x the x to set
-	 */
-	public void setX(Double x);
-
-	/**
-	 * @param y the y to set
-	 */
-	public void setY(Double y);
+	public void setId(String id);
 
 }
