+++ src/main/java/edu/ucsd/sbrg/escher/models/EscherCompartment.java
@@ -14,7 +14,7 @@
  * <http://www.gnu.org/licenses/lgpl-3.0-standalone.html>.
  * ---------------------------------------------------------------------
  */
-package edu.ucsd.sbrg.escher;
+package edu.ucsd.sbrg.escher.models;
 
 import java.text.MessageFormat;
 import java.util.ResourceBundle;
@@ -64,7 +64,7 @@ public class EscherCompartment extends AbstractBox implements Element {
 	}
 
 	/* (non-Javadoc)
-	 * @see edu.ucsd.sbrg.escher.AbstractEscherBase#clone()
+	 * @see edu.ucsd.sbrg.escher.models.AbstractEscherBase#clone()
 	 */
 	@Override
 	public EscherCompartment clone() {
@@ -72,7 +72,7 @@ public class EscherCompartment extends AbstractBox implements Element {
 	}
 
 	/* (non-Javadoc)
-	 * @see edu.ucsd.sbrg.escher.Element#getId()
+	 * @see edu.ucsd.sbrg.escher.models.Element#getId()
 	 */
 	@Override
 	public String getId() {
@@ -91,7 +91,7 @@ public class EscherCompartment extends AbstractBox implements Element {
 	}
 
 	/* (non-Javadoc)
-	 * @see edu.ucsd.sbrg.escher.Element#isSetId()
+	 * @see edu.ucsd.sbrg.escher.models.Element#isSetId()
 	 */
 	@Override
 	public boolean isSetId() {
@@ -107,7 +107,7 @@ public class EscherCompartment extends AbstractBox implements Element {
 	}
 
 	/* (non-Javadoc)
-	 * @see edu.ucsd.sbrg.escher.Element#setId(java.lang.String)
+	 * @see edu.ucsd.sbrg.escher.models.Element#setId(java.lang.String)
 	 */
 	@Override
 	public void setId(String id) {
